Refactor manager_rules.go using Strategy Pattern to improve maintainability and testability
