+++ .tasks/task-refactor-manager-rules.md
@@ -1,0 +1,29 @@
+# Task: Refactor manager_rules.go
+
+## Task Steps
+
+- [x] Step 1: Create a backup of the current manager_rules.go file
+- [x] Step 2: Create a new structure for handling different reference types
+- [x] Step 3: Extract lockfile update logic to a separate function
+- [x] Step 4: Remove goto statements with proper control flow
+- [x] Step 5: Create handler interfaces and implementations for different reference types
+- [x] Step 6: Refactor the addRuleByReferenceImpl function to use the handlers
+- [x] Step 7: Test the refactored code
+- [x] Step 8: Update documentation
+
+## Full Task Description
+
+Refactor the manager_rules.go file to improve its structure and remove goto statements. The main issues that need to be addressed are:
+
+1. The file is too large and contains complex control flow with goto statements
+2. The addRuleByReferenceImpl function has a large if/else chain to detect reference types
+3. There's potential duplication in local file handling logic
+
+The refactoring will focus on creating a cleaner structure with handler interfaces for different reference types, proper error handling, and removal of goto statements.
+
+## Considerations
+
+- Ensure backward compatibility
+- Maintain the same behavior for all reference types
+- Improve testability
+- Create more focused files by responsibility
