+++ pkg/manager/manager_github.go
@@ -157,16 +157,4 @@ func handleGitHubDir(cursorDir, ref string) (RuleSource, error) {
 	return RuleSource{}, fmt.Errorf("GitHub directory references are not yet supported: %s", ref)
 }
 
-// handleLocalFile handles a local file reference.
-func handleLocalFile(cursorDir, ref string, isAbs bool) (RuleSource, error) {
-	rule, err := processLocalFile(cursorDir, ref, isAbs)
-	if err != nil {
-		return RuleSource{}, err
-	}
-
-	// Keep the original reference
-	rule.Reference = ref
-
-	Debugf("handleLocalFile completed with rule key: '%s'", rule.Key)
-	return rule, nil
-}
+// handleLocalFile is now moved to manager_local_handlers.go
