+++ pkg/manager/manager_rules.go
@@ -8,7 +8,6 @@ import (
 	"strings"
 
 	"github.com/fireharp/cursor-rules/pkg/templates"
-	"github.com/gobwas/glob"
 )
 
 // Function type definitions for dependency injection and testing
@@ -82,343 +81,53 @@ func AddRuleByReference(cursorDir, ref string) error {
 }
 
 // addRuleByReferenceImpl is the implementation of AddRuleByReference.
+// This function has been refactored to use the Strategy Pattern instead of a large
+// if/else chain. It now uses a registry of handlers to find the appropriate handler
+// for the given reference type, and delegates the processing to that handler.
+// The original goto statement has been replaced with proper function extraction.
 func addRuleByReferenceImpl(cursorDir, ref string) error {
-	// Skip if already installed - we need to detect the key first
-	var rule RuleSource
-	var err error
+	// Create a registry with all reference handlers
+	registry := NewReferenceHandlerRegistry()
 
-	// Handle different reference types in order from most specific to least specific
-	if isGlobPattern(ref) {
-		// Handle glob pattern
-		return handleGlobPattern(context.Background(), cursorDir, ref)
-	} else if isGitHubBlobURL(ref) {
-		rule, err = handleGitHubBlob(context.Background(), cursorDir, ref)
-	} else if isGitHubTreeURL(ref) {
-		rule, err = handleGitHubDir(cursorDir, ref)
-	} else if isAbsolutePath(ref) {
-		// Check absolute paths before username patterns
-		rule, err = handleLocalFile(cursorDir, ref, true)
-	} else if isRelativePath(ref) {
-		// Check relative paths before username patterns
-		rule, err = handleLocalFile(cursorDir, ref, false)
-	} else if isUsernameRuleWithSha(ref) {
-		// Handle username/rule:sha format
-		rule, err = handleUsernameRuleWithSha(context.Background(), cursorDir, ref)
-	} else if isUsernameRuleWithTag(ref) {
-		// Handle username/rule@tag format
-		rule, err = handleUsernameRuleWithTag(context.Background(), cursorDir, ref)
-	} else if isUsernameRule(ref) {
-		// Handle username/rule format
-		rule, err = handleUsernameRule(context.Background(), cursorDir, ref)
-	} else if isUsernamePathRule(ref) {
-		// Handle username/path/rule format
-		rule, err = handleUsernamePathRule(context.Background(), cursorDir, ref)
-	} else {
-		// Check if there's a default username and this is a simple rule name
-		defaultUsername := getDefaultUsername()
-		if defaultUsername != "" {
-			// Use the default username for resolution
-			defaultRef := defaultUsername + "/" + ref
-			// Directly construct GitHub URL to avoid recursive call
-			githubURL := fmt.Sprintf("https://github.com/%s/cursor-rules-collection/blob/main/%s.mdc",
-				defaultUsername, ref)
-			rule, err = handleGitHubBlob(context.Background(), cursorDir, githubURL)
-
-			if err == nil {
-				// Found in the cursor-rules-collection repo
-				rule.SourceType = SourceTypeGitHubShorthand
-				rule.Reference = defaultRef // Store the resolved reference
-				goto checkLockfile
-			}
-
-			// If not found, try with potential paths (could be nested)
-			githubURL = fmt.Sprintf("https://github.com/%s/cursor-rules-collection/blob/main/%s/%s.mdc",
-				defaultUsername, ref, ref)
-			rule, err = handleGitHubBlob(context.Background(), cursorDir, githubURL)
-
-			if err == nil {
-				// Found in the cursor-rules-collection repo in a subdirectory
-				rule.SourceType = SourceTypeGitHubShorthand
-				rule.Reference = defaultRef // Store the resolved reference
-				goto checkLockfile
-			}
-		}
-
-		// Use the built-in template approach (search in templates)
+	// Find a handler for this reference
+	handler := registry.FindHandler(ref)
+	if handler == nil {
+		// Check if it's a built-in template as a last resort
 		tmpl, err := templates.FindTemplateByName(ref)
 		if err == nil && tmpl.Category != "" {
 			return AddRule(cursorDir, tmpl.Category, ref)
 		}
+
 		return fmt.Errorf("unsupported reference format or rule not found: %s", ref)
 	}
 
-	if err != nil {
-		return fmt.Errorf("failed to process reference: %w", err)
-	}
+	// Process the reference
+	rule, err := handler.Process(context.Background(), cursorDir, ref)
 
-checkLockfile:
-	// Check if rule is already installed
-	lock, err := LoadLockFile(cursorDir)
+	// Handle specific error cases
 	if err != nil {
-		return fmt.Errorf("failed to load lockfile: %w", err)
-	}
-
-	// Check if rule is already installed and handle gracefully
-	if lock.IsInstalled(rule.Key) {
-		// Find the existing rule
-		var existingRule RuleSource
-		for _, r := range lock.Rules {
-			if r.Key == rule.Key {
-				existingRule = r
-				break
+		// Check if this is a special error indicating a template was found
+		if strings.HasPrefix(err.Error(), "template_found:") {
+			parts := strings.Split(err.Error(), ":")
+			if len(parts) == 3 {
+				category := parts[1]
+				templateName := parts[2]
+				return AddRule(cursorDir, category, templateName)
 			}
 		}
 
-		// If this is a GitHub rule, check if commits differ
-		if rule.ResolvedCommit != "" && existingRule.ResolvedCommit != "" &&
-			rule.ResolvedCommit != existingRule.ResolvedCommit {
-			fmt.Printf("Rule '%s' is already installed but has a different version.\n", rule.Key)
-			fmt.Printf("  Current commit: %s\n", existingRule.ResolvedCommit)
-			fmt.Printf("  New commit: %s\n", rule.ResolvedCommit)
-			fmt.Printf("To update, use: cursor-rules upgrade %s\n", rule.Key)
-			return nil
-		}
+		// Pass through other errors
+		return fmt.Errorf("failed to process reference: %w", err)
+	}
 
-		// If content hash is available, compare that
-		if rule.ContentSHA256 != "" && existingRule.ContentSHA256 != "" &&
-			rule.ContentSHA256 != existingRule.ContentSHA256 {
-			fmt.Printf("Rule '%s' is already installed but has different content.\n", rule.Key)
-			fmt.Printf("To update, use: cursor-rules upgrade %s\n", rule.Key)
-			return nil
-		}
-
-		// If we get here, the rule is the same or we can't determine differences
-		fmt.Printf("Rule '%s' is already installed and up-to-date.\n", rule.Key)
+	// For glob patterns, the handler updates the lockfile directly and returns an empty RuleSource
+	if handler.CanHandle(ref) && isGlobPattern(ref) {
+		// Glob patterns are handled specially and already update the lockfile
 		return nil
 	}
 
-	// Update lockfile with the new rule
-	lock.Rules = append(lock.Rules, rule)
-	// For backwards compatibility
-	lock.Installed = append(lock.Installed, rule.Key)
-
-	err = lock.Save(cursorDir)
-	if err != nil {
-		return fmt.Errorf("failed to update lockfile: %w", err)
-	}
-
-	return nil
-}
-
-// handleUsernameRule handles a reference in the username/rule format.
-// This will look for the rule in the username/cursor-rules-collection repo.
-func handleUsernameRule(ctx context.Context, cursorDir, ref string) (RuleSource, error) {
-	username, ruleName, ok := parseUsernameRule(ref)
-	if !ok {
-		return RuleSource{}, fmt.Errorf("invalid username/rule format: %s", ref)
-	}
-
-	Debugf("handleUsernameRule: username='%s', ruleName='%s'\n", username, ruleName)
-
-	// Try to find it in username/cursor-rules-collection repo at root level only
-	githubURL := fmt.Sprintf("https://github.com/%s/cursor-rules-collection/blob/main/%s.mdc", username, ruleName)
-	Debugf("handleUsernameRule: trying URL: %s\n", githubURL)
-
-	rule, err := handleGitHubBlob(ctx, cursorDir, githubURL)
-
-	if err == nil {
-		// Found in the cursor-rules-collection repo
-		rule.SourceType = SourceTypeGitHubShorthand
-		rule.Reference = ref // Store the original reference
-		Debugf("handleUsernameRule: Found rule at primary URL\n")
-		return rule, nil
-	}
-
-	Debugf("handleUsernameRule: Primary URL failed with error: %v\n", err)
-
-	// No subfolder fallback for username/rule format
-	// This is intentional - username/rule should only look for the rule at the root level
-	// If users want a nested rule, they should use username/path/rule format
-
-	return RuleSource{}, fmt.Errorf("rule not found in username/cursor-rules-collection: %s", ref)
-}
-
-// handleUsernamePathRule handles a reference in the username/path/rule format.
-// This will first try in username/cursor-rules-collection, then fallback to username/repo.
-func handleUsernamePathRule(ctx context.Context, cursorDir, ref string) (RuleSource, error) {
-	username, pathParts, ok := parseUsernamePathRule(ref)
-	if !ok || len(pathParts) < 1 {
-		return RuleSource{}, fmt.Errorf("invalid username/path/rule format: %s", ref)
-	}
-
-	Debugf("handleUsernamePathRule: username='%s', pathParts=%v\n", username, pathParts)
-
-	// First, try to interpret it as username/path/to/rule in cursor-rules-collection
-	if len(pathParts) >= 1 {
-		// Extract the last part as the rule name and construct the path
-		ruleName := pathParts[len(pathParts)-1]
-		pathToRule := strings.Join(pathParts[:len(pathParts)-1], "/")
-
-		// If the rule already has .mdc extension, don't add it again
-		ruleFile := ruleName
-		if !strings.HasSuffix(ruleName, ".mdc") {
-			ruleFile = ruleName + ".mdc"
-		}
-
-		var githubURL string
-		if pathToRule == "" {
-			githubURL = fmt.Sprintf("https://github.com/%s/cursor-rules-collection/blob/main/%s",
-				username, ruleFile)
-		} else {
-			githubURL = fmt.Sprintf("https://github.com/%s/cursor-rules-collection/blob/main/%s/%s",
-				username, pathToRule, ruleFile)
-		}
-
-		Debugf("handleUsernamePathRule: trying URL: %s\n", githubURL)
-
-		rule, err := handleGitHubBlob(ctx, cursorDir, githubURL)
-		if err == nil {
-			// Found in the cursor-rules-collection repo
-			rule.SourceType = SourceTypeGitHubShorthand
-			rule.Reference = ref // Store the original reference
-			Debugf("handleUsernamePathRule: Found rule in cursor-rules-collection\n")
-			return rule, nil
-		}
-
-		Debugf("handleUsernamePathRule: cursor-rules-collection URL failed with error: %v\n", err)
-	}
-
-	// As a fallback, attempt to interpret it as username/repo/path/to/rule.mdc
-	if len(pathParts) >= 2 {
-		repoName := pathParts[0]
-		remainingPath := strings.Join(pathParts[1:], "/")
-
-		// If the last part already has .mdc extension, don't add it again
-		if !strings.HasSuffix(remainingPath, ".mdc") {
-			remainingPath += ".mdc"
-		}
-
-		githubURL := fmt.Sprintf("https://github.com/%s/%s/blob/main/%s",
-			username, repoName, remainingPath)
-
-		Debugf("handleUsernamePathRule: trying fallback URL (any repo): %s\n", githubURL)
-
-		rule, err := handleGitHubBlob(ctx, cursorDir, githubURL)
-		if err == nil {
-			// Found in the other repo
-			rule.SourceType = SourceTypeGitHubRepoPath
-			rule.Reference = ref // Store the original reference
-			Debugf("handleUsernamePathRule: Found rule in other repo\n")
-			return rule, nil
-		}
-
-		Debugf("handleUsernamePathRule: fallback URL failed with error: %v\n", err)
-	}
-
-	// Extra fallback: Try to handle the special case of "username/path/rule"
-	// This is specifically for fireharp/monorepo/monorepo type paths
-	if len(pathParts) >= 1 {
-		// Try to look for the file at "path/rule.mdc" in cursor-rules-collection
-		// This creates a URL like: https://github.com/fireharp/cursor-rules-collection/blob/main/monorepo/monorepo.mdc
-
-		// Join all path parts with /
-		fullPath := strings.Join(pathParts, "/")
-
-		// If the last part already has .mdc extension, don't add it again
-		if !strings.HasSuffix(fullPath, ".mdc") {
-			fullPath += ".mdc"
-		}
-
-		githubURL := fmt.Sprintf("https://github.com/%s/cursor-rules-collection/blob/main/%s",
-			username, fullPath)
-
-		Debugf("handleUsernamePathRule: trying second fallback URL (nested structure): %s\n", githubURL)
-
-		rule, err := handleGitHubBlob(ctx, cursorDir, githubURL)
-		if err == nil {
-			// Found in the cursor-rules-collection repo with the nested structure
-			rule.SourceType = SourceTypeGitHubShorthand
-			rule.Reference = ref // Store the original reference
-			Debugf("handleUsernamePathRule: Found rule in cursor-rules-collection nested structure\n")
-			return rule, nil
-		}
-
-		Debugf("handleUsernamePathRule: second fallback URL failed with error: %v\n", err)
-	}
-
-	return RuleSource{}, fmt.Errorf("rule not found in any repo: %s", ref)
-}
-
-// handleUsernameRuleWithSha handles a reference in the username/rule:sha format.
-func handleUsernameRuleWithSha(ctx context.Context, cursorDir, ref string) (RuleSource, error) {
-	username, ruleName, sha, ok := parseUsernameRuleWithSha(ref)
-	if !ok {
-		return RuleSource{}, fmt.Errorf("invalid username/rule:sha format: %s", ref)
-	}
-
-	// Build GitHub URL with specific commit
-	githubURL := fmt.Sprintf("https://github.com/%s/cursor-rules-collection/blob/%s/%s.mdc",
-		username, sha, ruleName)
-
-	rule, err := handleGitHubBlob(ctx, cursorDir, githubURL)
-	if err == nil {
-		// Found in the cursor-rules-collection repo at specific commit
-		rule.SourceType = SourceTypeGitHubShorthand
-		rule.Reference = ref // Store the original reference
-		rule.GitRef = "commit=" + sha
-		return rule, nil
-	}
-
-	// Try in a subdirectory as well
-	githubURL = fmt.Sprintf("https://github.com/%s/cursor-rules-collection/blob/%s/%s/%s.mdc",
-		username, sha, ruleName, ruleName)
-
-	rule, err = handleGitHubBlob(ctx, cursorDir, githubURL)
-	if err == nil {
-		// Found in the cursor-rules-collection repo in a subdirectory at specific commit
-		rule.SourceType = SourceTypeGitHubShorthand
-		rule.Reference = ref // Store the original reference
-		rule.GitRef = "commit=" + sha
-		return rule, nil
-	}
-
-	return RuleSource{}, fmt.Errorf("rule not found in username/cursor-rules-collection at commit %s: %s", sha, ref)
-}
-
-// handleUsernameRuleWithTag handles a reference in the username/rule@tag format.
-func handleUsernameRuleWithTag(ctx context.Context, cursorDir, ref string) (RuleSource, error) {
-	username, ruleName, tag, ok := parseUsernameRuleWithTag(ref)
-	if !ok {
-		return RuleSource{}, fmt.Errorf("invalid username/rule@tag format: %s", ref)
-	}
-
-	// Build GitHub URL with specific tag
-	githubURL := fmt.Sprintf("https://github.com/%s/cursor-rules-collection/blob/%s/%s.mdc",
-		username, tag, ruleName)
-
-	rule, err := handleGitHubBlob(ctx, cursorDir, githubURL)
-	if err == nil {
-		// Found in the cursor-rules-collection repo at specific tag
-		rule.SourceType = SourceTypeGitHubShorthand
-		rule.Reference = ref // Store the original reference
-		rule.GitRef = "tag=" + tag
-		return rule, nil
-	}
-
-	// Try in a subdirectory as well
-	githubURL = fmt.Sprintf("https://github.com/%s/cursor-rules-collection/blob/%s/%s/%s.mdc",
-		username, tag, ruleName, ruleName)
-
-	rule, err = handleGitHubBlob(ctx, cursorDir, githubURL)
-	if err == nil {
-		// Found in the cursor-rules-collection repo in a subdirectory at specific tag
-		rule.SourceType = SourceTypeGitHubShorthand
-		rule.Reference = ref // Store the original reference
-		rule.GitRef = "tag=" + tag
-		return rule, nil
-	}
-
-	return RuleSource{}, fmt.Errorf("rule not found in username/cursor-rules-collection at tag %s: %s", tag, ref)
+	// Update the lockfile with the processed rule
+	return updateLockfileWithRule(cursorDir, rule)
 }
 
 // RemoveRule uninstalls a rule and removes its files.
@@ -611,425 +320,5 @@ func SyncLocalRules(cursorDir string) error {
 		}
 	}
 
-	return nil
-}
-
-// handleGlobPattern handles references with glob patterns.
-// This processes multiple rules that match the pattern.
-func handleGlobPattern(ctx context.Context, cursorDir, ref string) error {
-	// Parse the glob pattern
-	username, pattern, ok := parseGlobPattern(ref)
-	if !ok {
-		return fmt.Errorf("invalid glob pattern: %s", ref)
-	}
-
-	// Compile the glob pattern
-	g, err := compileGlob(pattern)
-	if err != nil {
-		return fmt.Errorf("invalid glob pattern: %w", err)
-	}
-
-	// Check if the pattern is a local path (contains ./ or / at start, or no username)
-	if username == "" && (strings.HasPrefix(pattern, "./") || strings.HasPrefix(pattern, "/") || !strings.Contains(pattern, "/")) {
-		// Handle local file glob pattern
-		return handleLocalGlobPattern(ctx, cursorDir, pattern, g)
-	}
-
-	// Check if we have a username
-	if username != "" {
-		// Try to find matching rules in username/cursor-rules-collection
-		return handleUsernameGlobPattern(ctx, cursorDir, username, pattern, g)
-	} else {
-		// Try to find matching templates
-		return handleTemplateGlobPattern(cursorDir, pattern, g)
-	}
-}
-
-// handleLocalGlobPattern handles glob patterns for local filesystem.
-func handleLocalGlobPattern(ctx context.Context, cursorDir, pattern string, g glob.Glob) error {
-	Debugf("Processing local glob pattern: %s\n", pattern)
-
-	// Load lockfile once at the beginning
-	lock, err := LoadLockFile(cursorDir)
-	if err != nil {
-		return fmt.Errorf("failed to load lockfile: %w", err)
-	}
-
-	// Resolve pattern to absolute pattern if it's relative
-	var absolutePattern string
-	if strings.HasPrefix(pattern, "/") {
-		// Already absolute
-		absolutePattern = pattern
-	} else {
-		// Get current directory and join with pattern
-		cwd, err := os.Getwd()
-		if err != nil {
-			return fmt.Errorf("failed to get current working directory: %w", err)
-		}
-
-		// If pattern starts with ./, remove it
-		if strings.HasPrefix(pattern, "./") {
-			pattern = pattern[2:]
-		}
-
-		absolutePattern = filepath.Join(cwd, pattern)
-	}
-
-	Debugf("Using absolute glob pattern: %s\n", absolutePattern)
-
-	// Use filepath.Glob to expand the pattern
-	matchedFiles, err := filepath.Glob(absolutePattern)
-	if err != nil {
-		return fmt.Errorf("failed to expand glob pattern: %w", err)
-	}
-
-	Debugf("Found %d matching files\n", len(matchedFiles))
-
-	if len(matchedFiles) == 0 {
-		return fmt.Errorf("no files matched the pattern: %s", pattern)
-	}
-
-	// Track success/failure
-	successCount := 0
-	skippedCount := 0
-	errorCount := 0
-
-	// Collection of new rules
-	newRules := []RuleSource{}
-
-	// Process each matching file
-	for _, filePath := range matchedFiles {
-		// Skip directories
-		info, err := os.Stat(filePath)
-		if err != nil {
-			fmt.Printf("Warning: Could not stat file %s: %v\n", filePath, err)
-			errorCount++
-			continue
-		}
-
-		if info.IsDir() {
-			fmt.Printf("Skipping directory: %s\n", filePath)
-			continue
-		}
-
-		// Skip non-mdc files
-		if !strings.HasSuffix(filePath, ".mdc") {
-			fmt.Printf("Skipping non-mdc file: %s\n", filePath)
-			continue
-		}
-
-		fmt.Printf("Processing file: %s\n", filePath)
-
-		// Process the file - create a helper function that returns the RuleSource without updating lockfile
-		rule, err := processLocalFile(cursorDir, filePath, true)
-		if err != nil {
-			fmt.Printf("Warning: Could not process file %s: %v\n", filePath, err)
-			errorCount++
-			continue
-		}
-
-		// Check if rule is already installed
-		if lock.IsInstalled(rule.Key) {
-			// Find the existing rule
-			var existingRule RuleSource
-			for _, r := range lock.Rules {
-				if r.Key == rule.Key {
-					existingRule = r
-					break
-				}
-			}
-
-			// If content hash is available, compare that
-			if rule.ContentSHA256 != "" && existingRule.ContentSHA256 != "" &&
-				rule.ContentSHA256 != existingRule.ContentSHA256 {
-				fmt.Printf("Rule '%s' is already installed but has different content.\n", rule.Key)
-				fmt.Printf("To update, use: cursor-rules upgrade %s\n", rule.Key)
-			} else {
-				fmt.Printf("Rule '%s' is already installed and up-to-date.\n", rule.Key)
-			}
-
-			skippedCount++
-			continue
-		}
-
-		// Check if we've already added this rule in the current operation
-		alreadyAdded := false
-		for _, r := range newRules {
-			if r.Key == rule.Key {
-				alreadyAdded = true
-				break
-			}
-		}
-
-		if alreadyAdded {
-			fmt.Printf("Rule already added in this operation: %s\n", rule.Key)
-			continue
-		}
-
-		// Add rule to our collection
-		newRules = append(newRules, rule)
-		successCount++
-	}
-
-	// Update lockfile with all new rules
-	if len(newRules) > 0 {
-		lock.Rules = append(lock.Rules, newRules...)
-
-		// For backwards compatibility
-		for _, rule := range newRules {
-			lock.Installed = append(lock.Installed, rule.Key)
-		}
-
-		err = lock.Save(cursorDir)
-		if err != nil {
-			return fmt.Errorf("failed to update lockfile: %w", err)
-		}
-	}
-
-	// Report results
-	if successCount == 0 && skippedCount == 0 {
-		return fmt.Errorf("no valid rules found for pattern: %s", pattern)
-	}
-
-	fmt.Printf("Added %d rules matching pattern %s (skipped: %d, errors: %d)\n",
-		successCount, pattern, skippedCount, errorCount)
-	return nil
-}
-
-// processLocalFile processes a local file and returns a RuleSource without updating the lockfile.
-// This is a helper function extracted from handleLocalFile to be used with glob patterns.
-func processLocalFile(cursorDir, filePath string, isAbs bool) (RuleSource, error) {
-	// 1. Validate path and ensure it's readable
-	var fullPath string
-	if isAbs {
-		fullPath = filePath
-	} else {
-		var err error
-		fullPath, err = filepath.Abs(filePath)
-		if err != nil {
-			return RuleSource{}, fmt.Errorf("failed to resolve path %s: %w", filePath, err)
-		}
-	}
-
-	// Check if the file exists and is readable
-	info, err := os.Stat(fullPath)
-	if err != nil {
-		return RuleSource{}, fmt.Errorf("failed to access file %s: %w", fullPath, err)
-	}
-
-	if info.IsDir() {
-		return RuleSource{}, fmt.Errorf("%s is a directory, not a file", fullPath)
-	}
-
-	// 2. Read the file
-	data, err := os.ReadFile(fullPath)
-	if err != nil {
-		return RuleSource{}, fmt.Errorf("failed to read file %s: %w", fullPath, err)
-	}
-
-	// 3. Generate rule key and determine destination filename
-	ruleKey := generateRuleKey(filePath)
-	destFilename := ruleKey + ".mdc"
-	destPath := filepath.Join(cursorDir, destFilename)
-
-	// Ensure parent directories exist for hierarchical keys
-	if err := ensureRuleDirectory(cursorDir, ruleKey); err != nil {
-		return RuleSource{}, fmt.Errorf("failed preparing directory for rule '%s': %w", ruleKey, err)
-	}
-
-	// 4. Write to .cursor/rules
-	err = os.WriteFile(destPath, data, 0o600)
-	if err != nil {
-		return RuleSource{}, fmt.Errorf("failed to write to %s: %w", destPath, err)
-	}
-
-	// 5. Create and return RuleSource
-	sourceType := SourceTypeLocalAbs
-	if !isAbs {
-		sourceType = SourceTypeLocalRel
-		// Use relative path if possible for portability
-		// Get current working directory
-		cwd, err := os.Getwd()
-		if err == nil {
-			rel, err := filepath.Rel(cwd, fullPath)
-			if err == nil {
-				fullPath = rel
-			}
-		}
-	}
-
-	// Create the rule source
-	result := RuleSource{
-		Key:        ruleKey,
-		SourceType: sourceType,
-		Reference:  filePath,
-		LocalFiles: []string{destFilename},
-		// Calculate and store content hash for future modification checks
-		ContentSHA256: calculateSHA256(data),
-		// Store the original glob pattern
-		GlobPattern: filePath,
-	}
-
-	return result, nil
-}
-
-// handleUsernameGlobPattern handles glob patterns with a username.
-// This looks for matching rules in the username's cursor-rules-collection repo.
-func handleUsernameGlobPattern(ctx context.Context, cursorDir, username, pattern string, g glob.Glob) error {
-	// Try to get a list of files from the repo
-	owner := username
-	repo := "cursor-rules-collection"
-	branch := "main" // Default to main branch
-
-	// Get list of files from GitHub
-	files, err := listGitHubRepoFiles(ctx, owner, repo, branch, pattern)
-	if err != nil {
-		// If we can't list, tell the user and suggest alternatives
-		fmt.Printf("Could not list files matching pattern: %v\n", err)
-		fmt.Println("You may need to add rules individually instead of using glob patterns.")
-		return fmt.Errorf("failed to list files matching pattern: %w", err)
-	}
-
-	// Track success/failure
-	successCount := 0
-	errorCount := 0
-
-	// Process each matching file
-	for _, file := range files {
-		// Check if it matches our pattern
-		if !matchGlob(g, file) {
-			continue
-		}
-
-		// Only process .mdc files
-		if !strings.HasSuffix(file, ".mdc") {
-			continue
-		}
-
-		// Construct a new reference without the glob
-		fileRef := fmt.Sprintf("%s/%s", username, file)
-
-		// Directly construct GitHub URL to avoid recursive call to AddRuleByReference
-		githubURL := fmt.Sprintf("https://github.com/%s/cursor-rules-collection/blob/main/%s",
-			username, file)
-
-		rule, err := handleGitHubBlob(ctx, cursorDir, githubURL)
-		if err != nil {
-			fmt.Printf("Warning: Could not add rule %s: %v\n", fileRef, err)
-			errorCount++
-		} else {
-			// Found in the cursor-rules-collection repo
-			rule.SourceType = SourceTypeGitHubShorthand
-			rule.Reference = fileRef // Store the original reference
-
-			// Update lockfile with the new rule
-			lock, err := LoadLockFile(cursorDir)
-			if err != nil {
-				fmt.Printf("Warning: Could not load lockfile for %s: %v\n", fileRef, err)
-				errorCount++
-				continue
-			}
-
-			if lock.IsInstalled(rule.Key) {
-				fmt.Printf("Rule already installed: %s\n", rule.Key)
-				continue
-			}
-
-			lock.Rules = append(lock.Rules, rule)
-			// For backwards compatibility
-			lock.Installed = append(lock.Installed, rule.Key)
-
-			err = lock.Save(cursorDir)
-			if err != nil {
-				fmt.Printf("Warning: Could not update lockfile for %s: %v\n", fileRef, err)
-				errorCount++
-				continue
-			}
-
-			successCount++
-		}
-	}
-
-	// Report results
-	if successCount == 0 {
-		return fmt.Errorf("no matching rules found for pattern: %s", pattern)
-	}
-
-	fmt.Printf("Added %d rules matching pattern %s (errors: %d)\n", successCount, pattern, errorCount)
-	return nil
-}
-
-// handleTemplateGlobPattern handles glob patterns for built-in templates.
-func handleTemplateGlobPattern(cursorDir, pattern string, g glob.Glob) error {
-	// Get list of available templates
-	allTemplates, err := templates.ListTemplates()
-	if err != nil {
-		return fmt.Errorf("failed to list templates: %w", err)
-	}
-
-	// Track success/failure
-	successCount := 0
-	errorCount := 0
-
-	// Process each matching template
-	for _, tmpl := range allTemplates {
-		// Check if the template name matches our pattern
-		if !matchGlob(g, tmpl.Name) {
-			continue
-		}
-
-		// Try to add this template directly instead of calling AddRule
-		content := tmpl.Content
-		targetPath := filepath.Join(cursorDir, tmpl.Name+".mdc")
-
-		// Write to .cursor/rules/{ruleName}.mdc
-		err := os.WriteFile(targetPath, []byte(content), 0o644)
-		if err != nil {
-			fmt.Printf("Warning: Could not write template %s: %v\n", tmpl.Name, err)
-			errorCount++
-			continue
-		}
-
-		// Update lockfile
-		lock, err := LoadLockFile(cursorDir)
-		if err != nil {
-			fmt.Printf("Warning: Could not load lockfile for %s: %v\n", tmpl.Name, err)
-			errorCount++
-			continue
-		}
-
-		if lock.IsInstalled(tmpl.Name) {
-			fmt.Printf("Template already installed: %s\n", tmpl.Name)
-			continue
-		}
-
-		rule := RuleSource{
-			Key:        tmpl.Name,
-			SourceType: SourceTypeBuiltIn,
-			Reference:  tmpl.Name,
-			Category:   tmpl.Category,
-			LocalFiles: []string{targetPath},
-		}
-
-		lock.Rules = append(lock.Rules, rule)
-		// For backwards compatibility
-		lock.Installed = append(lock.Installed, tmpl.Name)
-
-		err = lock.Save(cursorDir)
-		if err != nil {
-			fmt.Printf("Warning: Could not update lockfile for %s: %v\n", tmpl.Name, err)
-			errorCount++
-			continue
-		}
-
-		successCount++
-	}
-
-	// Report results
-	if successCount == 0 {
-		return fmt.Errorf("no matching templates found for pattern: %s", pattern)
-	}
-
-	fmt.Printf("Added %d templates matching pattern %s (errors: %d)\n", successCount, pattern, errorCount)
 	return nil
 }
